+++ test/invariants/EthenaARM/TargetFunctions.sol
@@ -327,7 +327,7 @@ abstract contract TargetFunctions is Setup, StdUtils {
         crossPrice = _bound(crossPrice, minCrossPrice, maxCrossPrice);
 
         uint256 susdeBalance = susde.balanceOf(address(arm));
-        (,,,,, uint120 pendingRedeemAssets,,,) = arm.baseAssetConfigs(address(susde));
+        (,,,,, uint128 pendingRedeemAssets,,,) = arm.baseAssetConfigs(address(susde));
         bool loweringCrossPrice = _crossPrice() > crossPrice;
         if (loweringCrossPrice && assume(uint256(pendingRedeemAssets) < DEFAULT_MIN_TOTAL_SUPPLY)) return;
